src/requirements.txt
Flask-JWT-Extended
Flask-Smorest
Flask-SQLAlchemy
marshmallow
python-dotenv
passlib
requests